+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-12T12:33:09.842Z",
+    "2026-05-12T15:17:11.138Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {