+++ packages/orq-rc/RELEASES.md
@@ -9109,3 +9109,13 @@ Based on:
 - [typescript v4.9.0-rc.50] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.50 - packages/orq-rc
+
+## 2026-05-12 15:20:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.51] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.51] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.51 - packages/orq-rc